refactor(js/plugins/vertexai)!: extract model garden into sub-plugin
