+++ js/plugins/vertexai/src/model-garden/anthropic.ts
