+++ js/plugins/vertexai/src/model-garden/model_garden.ts
