+++ js/plugins/vertexai/src/model-garden/openai_compatibility.ts
